refactor(城市系统): 修改城市属性类型为float并统一使用GetAttr方法
将城市属性(gold,food,soldier等)从int改为float类型
所有直接访问属性的地方改为使用GetAttr方法获取
调整城市发展和等级配置的数值平衡
更新UI显示和AI逻辑以适配新的属性类型

+++ Assets/Resources/Scripts/PopArmySetManager.cs
@@ -78,7 +78,7 @@ public class PopArmySetManager : MonoBehaviour
         heroPic.sprite = Resources.Load<Sprite>("Textures/Skins/" + heroCfg.Icon);
         textHeroName.text = heroCfg.Name;
 
-        maxSoldier = Math.Min(1000, GameManager.Instance.GetCity(cityId).soldier + heroData.soldier);
+        maxSoldier = Math.Min(1000, (int)(GameManager.Instance.GetCity(cityId).soldier) + heroData.soldier);
         slider1.value = (float)heroData.soldier / maxSoldier;
     }
